Fixed some warnings on Visual Studio 2019

+++ Lib-ZeroG/src/ZeroG/d3d12/D3D12CommandList.cpp
@@ -398,7 +398,7 @@ ZgErrorCode D3D12CommandList::setPipelineBindings(
 		srvDesc.ViewDimension = D3D12_SRV_DIMENSION_TEXTURE2D;
 		srvDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
 		srvDesc.Texture2D.MostDetailedMip = 0;
-		srvDesc.Texture2D.MipLevels = -1; // All mip-levels from most detailed and downwards
+		srvDesc.Texture2D.MipLevels = (uint32_t)-1; // All mip-levels from most detailed and downwards
 		srvDesc.Texture2D.PlaneSlice = 0;
 		srvDesc.Texture2D.ResourceMinLODClamp = 0.0f;
 		mDevice->CreateShaderResourceView(texture->resource.Get(), &srvDesc, cpuDescriptor);
